feat: new file for removing word commands
